+++ apps/integration/integration/ingests/index_info.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class IndexInfo:
+    name: str
+    num_docs: int
